FFM-9439: Allow feature flag to add rules to multiple environments (#742)
After this change, we will allow the user to deploy updates to multiple
environments. Previously, we only allowed one environment to be changed
per feature flag. Now, we let multiple environments to be changed.
Target Groups and Rules are now defined inside the environment

Also include document and examples update, as well as the changelog